adds offer data for offer page
