+++ project/src/components/index.ts
@@ -16,4 +16,5 @@ export { default as SortingForm } from './sorting-form/sorting-form';
 export { default as Spinner } from './spinner/spinner';
 export { default as NoLoggedNav } from './no-logged-nav/no-logged-nav';
 export { default as LoggedNav } from './logged-nav/logged-nav';
+export { default as OfferItem } from './offer-item/offer-item';
